+++ .vscode/settings.json
@@ -22,6 +22,7 @@
         "Dumouchelle",
         "openmarch",
         "Rowid",
+        "Topbar",
         "zustand"
     ],
     "[typescriptreact]": {
