+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## Features
 
-- None yet
+- Lines
+  - Users can select a group of 3 or more marchers and draw a straight line
+  - This is the start of a new feature called an "Alignment Event." This will be how other shapes are created as well
 
 ## Quality of life
 
@@ -14,3 +16,5 @@
 ## Dev
 
 - Switched from Jest to Vitest
+- Reorganized files for the sidebar, topbar, and timeline to have their own folders
+  - Sidebar editors now have their own container component `EditorContainer.tsx`
