+++ src/components/canvas/listeners/CanvasListeners.ts
@@ -6,4 +6,6 @@ export default interface CanvasListeners {
     initiateListeners: () => void;
     /** Use this function to detach listeners and perform any cleanup actions when switching the cursor mode */
     cleanupListeners: () => void;
+    /** A function to call when the canvas refreshes the marchers */
+    refreshMarchers?: () => void;
 }
