+++ src/components/music/MusicModal.tsx
@@ -1,6 +1,6 @@
 // import ModalLauncher from "../toolbar/ModalLauncher";
 import { topBarComponentProps } from "@/global/Interfaces";
-import ModalLauncher from "../toolbar/ModalLauncher";
+import ModalLauncher from "../topbar/ModalLauncher";
 import { FaMusic } from "react-icons/fa";
 import AudioSelector from "./AudioSelector";
 import { RegisteredActionsObjects } from "@/utilities/RegisteredActionsHandler";
