+++ src/components/page/PageListModal.tsx
@@ -1,4 +1,4 @@
-import ModalLauncher from "../toolbar/ModalLauncher";
+import ModalLauncher from "../topbar/ModalLauncher";
 import { useEffect, useState } from "react";
 import FormButtons from "../FormButtons";
 import PageList from "./PageList";
